refactor: extract accordion and accordion-panel mixins
